replace quorum List by Informer
List() send on request to the apiserver each time and it goes directly
to etcd.
Informers implement a reflector that first List using resourceVersion=0
to leverage the apiserver cache and avoid going to etcd.
Once the List is obtained it starts watching for incremental changes.
These changes are stored in a local cache that users can consume.
The informers also expose callbacks that for all type of changes
Add, Update or Delete