add workqueue
Adding a workqueue allow us to implement the controller pattern, the
events are preprocessed on the handlers and enqueued to be processed by
the workers.

The workqueue guarantee deduplication so multiple workers can not work
in the same item at the same time.

One of the common mistakes is to operate by events instead of by levels.