log time to process each items
one of the common problems is to be able to dimension the system
accordenly or have bugs or slow paths in the reconcile logic.
If all the workers are busy, no new items are consumed and the memory
start to grow since the watch fifo queue can not be drained.